+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.72.0
 
+### Changes
+- use webxdc app name in chatlist/quotes/replies etc. #3027
+
 ### Fixes
 - run migrations on backup import #3006